Made it so no ugly rerender with github posts. Tested with good github url, bad github url, and no github url
