+++ frontend/src/components/StreamFeed.jsx
@@ -7,7 +7,9 @@ import PropTypes from 'prop-types';
 import CreatePostModal from '../components/CreatePostModal';
 import { toast } from 'react-semantic-toasts';
 import 'react-semantic-toasts/styles/react-semantic-alert.css';
+import utils from "../util/utils";
 import './styles/StreamFeed.css';
+import Cookies from 'js-cookie';
 
 const controller = new AbortController();
 const signal = controller.signal;
@@ -21,8 +23,11 @@ class StreamFeed extends Component {
 			isFetching: false,
 			showModal: false,
 			github: [],
+			githuburl: '',
 		};
 		this.getPosts = this.getPosts.bind(this);
+		this.fetchProfile = this.fetchProfile.bind(this);
+		this.getloggedinAuthorID = this.getloggedinAuthorID.bind(this);
 		this.closeModal = this.closeModal.bind(this);
 		this.createPostFromJson = this.createPostFromJson.bind(this);
 		this.deletePost = this.deletePost.bind(this);
@@ -65,74 +70,106 @@ class StreamFeed extends Component {
 	};
 	
 	componentDidMount() {
-		if (this.props.githuburl) {
-			this.getGithub();
+		if (this.props.getGithub) {
+			this.fetchProfile();
 		}
 		else {
 			this.getPosts();
 		}
 	}
 	
-	componentDidUpdate(prevProps){
-		if(prevProps.githuburl !== this.props.githuburl){    
-			if (this.props.githuburl) {
-			this.getGithub();
-			}
+	getGithub() {
+		if (this.state.githuburl) {
+			const gituser = this.state.githuburl.split('/').filter(el => el).pop();
+			const gitUrl = "https://api.github.com/users/" + gituser + "/received_events/public";
+			let gitRequest = new Request(gitUrl);
+			fetch(gitRequest) 
+					.then(response => {
+							if (response.status === 200) {
+									response.json().then((results) =>  {
+											let eventarray = [];
+											for (let i = 0; i < results.length; i++) {
+													const type = results[i].type.split(/(?=[A-Z])/)
+													type.pop();
+										
+													const event = {
+															id: "G " + results[i].id,
+															profilePicture: null,
+															published: results[i].created_at,
+															title: results[i].type,
+															description: results[i].actor.display_login + " did " + results[i].type + " at " + results[i].repo.name,
+															content: "",
+															contentType: "text/plain",
+															categories: [],
+															visibility: "PUBLIC",
+															visibleTo: [],
+															unlisted: false,
+															author: {
+																id: "G " + results[i].id,
+																displayName: results[i].actor.display_login,
+															},
+															
+															origin: "https://github.com/",
+															
+															deletePost: null,
+															getPosts: null,
+															isGithub: true,
+													};
+													eventarray.push(event)
+											}
+											
+											this.setState({
+												github: eventarray
+											});
+											this.getPosts();
+									})
+							} else if (response.status === 404) {
+								this.getPosts();
+							} else {
+								throw new Error('Github API server rate limit exceeded!');
+							}
+					})
+					.catch(error => {
+							console.error(error);
+					});
+		}
+		else {
+			this.getPosts();
 		}
 	}
-	
 
-	getGithub() {
-		const gituser = this.props.githuburl.split('/').filter(el => el).pop();
-		const gitUrl = "https://api.github.com/users/" + gituser + "/received_events/public";
-		let gitRequest = new Request(gitUrl);
-		fetch(gitRequest) 
-				.then(response => {
-						if (response.status === 200) {
-								response.json().then((results) =>  {
-										let eventarray = [];
-										for (let i = 0; i < results.length; i++) {
-												const type = results[i].type.split(/(?=[A-Z])/)
-												type.pop();
-									
-												const event = {
-														id: "G " + results[i].id,
-														profilePicture: null,
-														published: results[i].created_at,
-														title: results[i].type,
-														description: results[i].actor.display_login + " did " + results[i].type + " at " + results[i].repo.name,
-														content: "",
-														contentType: "text/plain",
-														categories: [],
-														visibility: "PUBLIC",
-														visibleTo: [],
-														unlisted: false,
-														author: {
-															id: "G " + results[i].id,
-															displayName: results[i].actor.display_login,
-														},
-														
-														origin: "https://github.com/",
-														
-														deletePost: null,
-														getPosts: null,
-														isGithub: true,
-												};
-												eventarray.push(event)
-										}
-										
-										this.setState({
-											github: eventarray
-										});
-										this.getPosts();
+    getloggedinAuthorID() {
+        const cookieauthorid = Cookies.get("userID"),
+            storeauthorid = this.props.storeItems.userID;
+        let authorID;
+        if (cookieauthorid !== null) {
+            authorID = cookieauthorid;
+        } else if (storeauthorid !== null) {
+			authorID = storeauthorid;
+		}
+	return authorID;
+	}
+
+	fetchProfile() {
+		const authorID = this.getloggedinAuthorID();
+		const hostUrl = "/api/author/"+ utils.getShortAuthorId(authorID),
+				requireAuth = true;
+		HTTPFetchUtil.getRequest(hostUrl, requireAuth, signal)
+				.then((httpResponse) => {
+						if (httpResponse.status === 200) {
+								httpResponse.json().then((results) => {
+									this.setState({
+										githuburl: results.github,
+									});
+									this.getGithub();
 								})
 						} else {
-							throw new Error('Github API server rate limit exceeded!');
+							throw new Error('Could not get github username');
 						}
 				})
-				.catch(error => {
+				.catch((error) => {
 						console.error(error);
-				});
+		});
 	}
 
 	componentWillUnmount() {
@@ -152,7 +189,7 @@ class StreamFeed extends Component {
 						var postList = [];
 						
 						
-						if (this.props.githuburl && this.state.github){
+						if (this.props.getGithub && this.state.github){
 							let sortArray = this.state.github.concat(results.posts);
 							sortArray.sort(function(a, b) {
 							return (a.published > b.published) ? -1 : ((a.published < b.published) ? 1 : 0);
@@ -254,7 +291,7 @@ StreamFeed.defaultProps = {
 StreamFeed.propTypes = {
 	urlPath: PropTypes.string.isRequired,
 	storeItems: PropTypes.object.isRequired,
-	githuburl: PropTypes.string,
+	getGithub: PropTypes.bool,
 }
 
 export default StreamFeed;
